# Milky Way 2.0/retail_app/docs/developer/reference/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Milky Way 2.0 · Retail Agent",
    "version": "2.1.1"
  },
  "paths": {
...
    "/api/feedback": {
      "post": {
        "summary": "Feedback",
        "operationId": "feedback_api_feedback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Feedback"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/workspace/assets": {
      "get": {
        "tags": [
          "Improve workspace"
        ],
        "summary": "List Assets",
        "operationId": "list_assets_api_workspace_assets_get",
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 4000,
              "default": "",
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Improve workspace"
        ],
        "summary": "Create Asset",
        "operationId": "create_asset_api_workspace_assets_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspace/assets/{identity}": {
      "get": {
        "tags": [
          "Improve workspace"
        ],
        "summary": "Get Asset",
        "operationId": "get_asset_api_workspace_assets__identity__get",
        "parameters": [
          {
            "name": "identity",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identity"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Improve workspace"
        ],
        "summary": "Update Asset",
        "operationId": "update_asset_api_workspace_assets__identity__put",
        "parameters": [
          {
            "name": "identity",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identity"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspace/assets/{identity}/versions": {
      "get": {
        "tags": [
          "Improve workspace"
        ],
        "summary": "Versions",
        "operationId": "versions_api_workspace_assets__identity__versions_get",
        "parameters": [
          {
            "name": "identity",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identity"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspace/assets/{identity}/validate": {
      "post": {
        "tags": [
          "Improve workspace"
        ],
        "summary": "Validate Asset",
        "operationId": "validate_asset_api_workspace_assets__identity__validate_post",
        "parameters": [
          {
            "name": "identity",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identity"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspace/preview": {
      "post": {
        "tags": [
          "Improve workspace"
        ],
        "summary": "Preview",
        "operationId": "preview_api_workspace_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Preview"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspace/sources/{asset_id}/{version_id}": {
      "get": {
        "tags": [
          "Improve workspace"
        ],
        "summary": "Source",
        "operationId": "source_api_workspace_sources__asset_id___version_id__get",
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Asset Id"
            }
          },
          {
            "name": "version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Version Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspace/releases": {
      "get": {
        "tags": [
          "Improve workspace"
        ],
        "summary": "Releases",
        "operationId": "releases_api_workspace_releases_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Improve workspace"
        ],
        "summary": "Candidate",
        "operationId": "candidate_api_workspace_releases_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Candidate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspace/releases/{identity}": {
      "get": {
        "tags": [
          "Improve workspace"
        ],
        "summary": "Release",
        "operationId": "release_api_workspace_releases__identity__get",
        "parameters": [
          {
            "name": "identity",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identity"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspace/releases/{identity}/publish": {
      "post": {
        "tags": [
          "Improve workspace"
        ],
        "summary": "Publish",
        "operationId": "publish_api_workspace_releases__identity__publish_post",
        "parameters": [
          {
            "name": "identity",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identity"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Activation"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspace/releases/{identity}/rollback": {
      "post": {
        "tags": [
          "Improve workspace"
        ],
        "summary": "Rollback",
        "operationId": "rollback_api_workspace_releases__identity__rollback_post",
        "parameters": [
          {
            "name": "identity",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identity"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Activation"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/workspace/feedback": {
      "get": {
        "tags": [
          "Improve workspace"
        ],
        "summary": "Feedback",
        "operationId": "feedback_api_workspace_feedback_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Improve workspace"
        ],
        "summary": "Create Feedback",
        "operationId": "create_feedback_api_workspace_feedback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspace/feedback/{identity}": {
      "put": {
        "tags": [
          "Improve workspace"
        ],
        "summary": "Update Feedback",
        "operationId": "update_feedback_api_workspace_feedback__identity__put",
        "parameters": [
          {
            "name": "identity",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identity"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspace/feedback/{identity}/convert": {
      "post": {
        "tags": [
          "Improve workspace"
        ],
        "summary": "Convert Feedback",
        "operationId": "convert_feedback_api_workspace_feedback__identity__convert_post",
        "parameters": [
          {
            "name": "identity",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identity"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackConvert"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspace/output-preview": {
      "post": {
        "tags": [
          "Improve workspace"
        ],
        "summary": "Output Preview",
        "operationId": "output_preview_api_workspace_output_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OutputPreview"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Activation": {
        "properties": {
          "expected_active_release_id": {
            "type": "string",
            "title": "Expected Active Release Id"
          },
          "operator": {
            "type": "string",
            "maxLength": 100,
            "title": "Operator",
            "default": "local-admin"
          },
          "rationale": {
            "type": "string",
            "maxLength": 4000,
            "title": "Rationale",
            "default": ""
          },
          "evaluation_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evaluation Run Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_active_release_id"
        ],
        "title": "Activation"
      },
...
      "AssetCreate": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Name"
          },
          "content": {
            "additionalProperties": true,
            "type": "object",
            "title": "Content"
          },
          "owner": {
            "type": "string",
            "maxLength": 100,
            "title": "Owner",
            "default": "local-admin"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "name",
          "content"
        ],
        "title": "AssetCreate"
      },
      "AssetUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 160,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "content": {
            "additionalProperties": true,
            "type": "object",
            "title": "Content"
          },
          "expected_revision": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Expected Revision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "content",
          "expected_revision"
        ],
        "title": "AssetUpdate"
      },
      "Candidate": {
        "properties": {
          "asset_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Ids"
          },
          "name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Name",
            "default": "Workspace candidate"
          },
          "rationale": {
            "type": "string",
            "maxLength": 4000,
            "title": "Rationale",
            "default": ""
          },
          "operator": {
            "type": "string",
            "maxLength": 100,
            "title": "Operator",
            "default": "local-admin"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Candidate"
      },
...
      "Feedback": {
        "properties": {
          "analysis_id": {
            "type": "string",
            "title": "Analysis Id"
          },
          "rating": {
            "type": "string",
            "title": "Rating"
          },
          "comment": {
            "type": "string",
            "maxLength": 2000,
            "title": "Comment",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "analysis_id",
          "rating"
        ],
        "title": "Feedback"
      },
      "FeedbackConvert": {
        "properties": {
          "expected_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Revision"
          },
          "suite_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suite Id"
          },
          "case": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_revision"
        ],
        "title": "FeedbackConvert"
      },
      "FeedbackCreate": {
        "properties": {
          "analysis_id": {
            "type": "string",
            "title": "Analysis Id"
          },
          "issue_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 7,
            "title": "Issue Types"
          },
          "correction": {
            "type": "string",
            "maxLength": 8000,
            "title": "Correction",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "analysis_id"
        ],
        "title": "FeedbackCreate"
      },
      "FeedbackUpdate": {
        "properties": {
          "expected_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Revision"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "correction": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 8000
              },
              {
                "type": "null"
              }
            ],
            "title": "Correction"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_revision",
          "status"
        ],
        "title": "FeedbackUpdate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Preview": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 4000,
            "title": "Query",
            "default": ""
          },
          "asset_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Ids"
          },
          "limit": {
            "type": "integer",
            "maximum": 12.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 6
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Preview"
      },
...
    }
  }
}